Move PopoverRow to Popover folder
