+++ src/components/Calendar/CalendarDayPopover.vue
@@ -29,7 +29,7 @@
 import { useCalendar } from '../../use/calendar';
 import { CalendarDay } from '../../utils/page';
 import Popover from '../Popover/Popover.vue';
-import PopoverRow from '../PopoverRow/PopoverRow.vue';
+import PopoverRow from '../Popover/PopoverRow.vue';
 import CalendarSlot from './CalendarSlot.vue';
 
 const { dayPopoverId, displayMode, color, masks, locale } = useCalendar();
