+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export { default as Calendar } from './Calendar/Calendar.vue';
 export { default as DatePicker } from './DatePicker/DatePicker.vue';
 export { default as Popover } from './Popover/Popover.vue';
-export { default as PopoverRow } from './PopoverRow/PopoverRow.vue';
+export { default as PopoverRow } from './Popover/PopoverRow.vue';
